auth fragment - for authentication and as entry point on nav_graph.xml + refactored PrefsDataStore, as it is being used in multiple class across multiple files
now without proper authentication app can't be used, as well as even the values from different views can't be seen
