+++ app/src/main/java/dev/spikeysanju/expensetracker/data/local/datastore/PrefsDataStore.kt
@@ -1,0 +1,11 @@
+package dev.spikeysanju.expensetracker.data.local.datastore
+
+import android.content.Context
+import androidx.datastore.core.DataStore
+import androidx.datastore.preferences.core.Preferences
+import androidx.datastore.preferences.createDataStore
+
+
+abstract class PrefsDataStore(context: Context, fileName: String) {
+    internal val dataStore: DataStore<Preferences> = context.createDataStore(fileName)
+}
