+++ app/src/main/res/navigation/nav_graph.xml
@@ -3,7 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/nav_graph"
-    app:startDestination="@id/dashboardFragment">
+    app:startDestination="@id/authFragment">
 
     <fragment
         android:id="@+id/dashboardFragment"
@@ -91,6 +91,20 @@
         android:name="dev.spikeysanju.expensetracker.view.settings.SettingsFragment"
         android:label="Settings"
         tools:layout="@layout/fragment_settings" />
+
+    <fragment
+        android:id="@+id/authFragment"
+        android:name="dev.spikeysanju.expensetracker.view.auth.AuthFragment"
+        android:label="Authentication"
+        tools:layout="@layout/fragment_auth" >
+        <action
+            android:id="@+id/action_authFragment_to_dashboardFragment"
+            app:destination="@id/dashboardFragment"
+            app:enterAnim="@anim/slide_in_right"
+            app:exitAnim="@anim/slide_out_left"
+            app:popEnterAnim="@anim/slide_in_left"
+            app:popExitAnim="@anim/slide_out_right" />
+    </fragment>
     <dialog
         android:id="@+id/errorDialog"
         android:name="dev.spikeysanju.expensetracker.view.dialog.ErrorDialog"
